moving files around
https://coveord.atlassian.net/browse/KIT-1861
