+++ packages/atomic/src/components/common/result-list/table-display-results.tsx
