+++ packages/atomic/src/components/search/search-box-suggestions/atomic-search-box-instant-results/atomic-search-box-instant-results.tsx
@@ -16,7 +16,7 @@ import {
   BaseResultList,
   ResultListCommon,
   ResultRenderingFunction,
-} from '../../result-lists/result-list-common';
+} from '../../../common/result-list/result-list-common';
 import {
   ResultDisplayDensity,
   ResultDisplayImageSize,
